Deprecate `BeanStream.map` with `Fn` parameter, migrate to lambda-based alternative, and adjust serialization handling.
